+++ src/wintab.c
@@ -165,14 +165,16 @@ static LRESULT CALLBACK
 container_proc(HWND hwnd, UINT msg, WPARAM wp, LPARAM lp)
 {
   //printf("tabbar con_proc %03X\n", msg);
-static WORD dragmsg = 0;
 static int dragidx = -1;
-static int targidx = -1;
+static int dragini = -1;  // initial drag tab, where it was clicked on
 static int targpro = -1;
+static int xpos = 0;
+static HCURSOR hcursor = NULL;
 
   if (msg == WM_MOUSEACTIVATE) {
     //printf("WM_MOUSEACTIVATE lo %02X hi %02X\n", LOWORD(lp), HIWORD(lp));
     if (LOWORD(lp) == HTCLIENT && HIWORD(lp) == WM_LBUTTONDOWN) {
+      WORD dragmsg = 0;
       // begin drag-and-drop tab reordering
 #ifdef determine_tab_index_by_wnd_in_item
       // get tab wnd from TCITEM lParam, lookup tab index (not implemented)
@@ -192,6 +194,7 @@ static int targpro = -1;
       POINT p;
       if (GetCursorPos(&p) && ScreenToClient(hwnd, &p)) {
         int x = p.x - xoff;
+        xpos = p.x;
         dragidx = x / curr_tab_width;
         if (dragidx < ntabinfo)
           dragmsg = HIWORD(lp);
@@ -199,80 +202,61 @@ static int targpro = -1;
           dragidx = -1;
         //printf("%d:%d (pw %d) x %d drag %d\n", (int)p.y, (int)p.x, curr_tab_width, x, dragidx);
       }
+      dragini = dragidx; // suppress highlighting on initial drag position
+      if (dragmsg) {
+        SetCapture(hwnd);
+        hcursor = GetCursor();
+      }
     }
   }
-  else if (msg == WM_SETCURSOR && dragmsg && LOWORD(lp) == HTCLIENT) {
-    //printf("WM_SETCURSOR lo %02X hi %02X\n", LOWORD(lp), HIWORD(lp));
-    if (HIWORD(lp) == WM_MOUSEMOVE) {
-      // drag during tab reordering; display visual feedback
-
-      // cannot determine current cursor target tab via TCITEM
-      // (like #ifdef determine_tab_index_by_wnd_in_item above)
-      // as the index remains the click-and-drag tab
-
-      // enquire cursor position; derive drop target item index
+  else if (msg == WM_MOUSEMOVE) {
+    if ((GetCapture() == hwnd) && ((wp & MK_LBUTTON) != 0)) {
+      // dragging tab
       POINT p;
       if (GetCursorPos(&p) && ScreenToClient(hwnd, &p)) {
-        int x = p.x - xoff;
-        int dropidx = x / curr_tab_width;
-        //printf("%d:%d (pw %d) x %d drop %d\n", (int)p.y, (int)p.x, curr_tab_width, x, dropidx);
-        bool redraw_tab = false;
-        if (dropidx != targidx && targidx >= 0) {
-          // clear hover highlighting of previous drop target
+        if (abs(p.x - xpos) > GetSystemMetrics(SM_CXDRAG)) {
+          SetCursor(LoadCursor(NULL, IDC_SIZEWE));
 
-          // clear visual indication via RedrawWindow/WM_DRAWITEM
-          targidx = -1;
-          redraw_tab = true;
-        }
-        if (dropidx != dragidx) {
-          // visual indication of tab dragging: hover highlighting
-          // enquire tab rect to calculate relative cursor position
-          RECT tr;
-          SendMessage(tab_wnd, TCM_GETITEMRECT, dropidx, (LPARAM)&tr);
-          //printf("drop RECT %d %d %d %d\n", tr.left, tr.right, tr.top, tr.bottom);
-          int w = tr.right - tr.left;
-          int c = (tr.left + tr.right) / 2;
-          targpro = 100 - abs(x - c) * 100 / (w / 2);
-          //printf("drop %d [%d] %d: %d\n", tr.left, x, tr.right, targpro);
-#ifdef hover_tab_within_wm_setcursor
-#warning the RECT does not properly match what we have at WM_DRAWITEM ...
-          tr.right += xoff;  // tune tab position
-          tr.bottom += xoff;  // assuming yoff would be == xoff
-          HDC tabdc = GetDC(tab_wnd);
-          HBRUSH tabbr = CreateSolidBrush(RGB(50, 90, 200));
-          FillRect(tabdc, &tr, tabbr);
-          DeleteObject(tabbr);
-          ReleaseDC(tab_wnd, tabdc);
-#endif
-          // enquire tab rect to calculate relative cursor position
+          // drop tab while dragging for tab reordering
 
-          // render visual indication via RedrawWindow/WM_DRAWITEM
-          targidx = dropidx;
-          redraw_tab = true;
+          // derive drop target item index
+          int x = p.x - xoff;
+          int dropidx = x / curr_tab_width;
+          //printf("%d:%d (pw %d) x %d: drag %d -> drop %d\n", (int)p.y, (int)p.x, curr_tab_width, x, dragidx, dropidx);
+
+          // act on drop target item
+          if (dropidx < ntabinfo && dropidx >= 0) {
+            // visual indication of tab dragging: highlight the dragging
+            RECT tr; // enquire tab rect
+            SendMessage(tab_wnd, TCM_GETITEMRECT, dropidx, (LPARAM)&tr);
+            //printf("drop RECT %d %d %d %d\n", tr.left, tr.right, tr.top, tr.bottom);
+            // calculate relative cursor position
+            int w = tr.right - tr.left;
+            int c = (tr.left + tr.right) / 2;
+            targpro = 100 - abs(x - c) * 100 / (w / 2);
+            //printf("drop %d [%d] %d: %d%%\n", tr.left, x, tr.right, targpro);
+            // draw tab drag indication
+            RedrawWindow(tab_wnd, 0, 0, RDW_INVALIDATE | RDW_UPDATENOW);
+
+            // move the tab
+            win_tab_move(dropidx - dragidx);
+
+            // update index of tab being dragged
+            //printf("dragini %d dragidx %d dropidx %d\n", dragini, dragidx, dropidx);
+            dragidx = dropidx;
+            if (dragidx != dragini)
+              // enable dynamic highlighting when changed drag tab position
+              dragini = -1;
+          }
         }
-        if (redraw_tab)
-          RedrawWindow(tab_wnd, 0, 0, RDW_INVALIDATE | RDW_UPDATENOW);
       }
     }
-    else if (HIWORD(lp) == WM_LBUTTONUP) {
-      // drop tab while dragging for tab reordering
-
-      // cannot determine current cursor target tab via TCITEM
-      // (like #ifdef determine_tab_index_by_wnd_in_item above)
-      // as the index remains the click-and-drag tab
-
-      // enquire cursor position; derive drop target item index
-      POINT p;
-      if (GetCursorPos(&p) && ScreenToClient(hwnd, &p)) {
-        int x = p.x - xoff;
-        int dropidx = x / curr_tab_width;
-        //printf("%d:%d (pw %d) x %d: drag %d -> drop %d\n", (int)p.y, (int)p.x, curr_tab_width, x, dragidx, dropidx);
-
-        // act on drop target item
-        win_tab_move(dropidx - dragidx);
-        dragmsg = 0;
-      }
-    }
+  }
+  else if (msg == WM_LBUTTONUP && GetCapture() == hwnd) {
+    SetCursor(hcursor);
+    ReleaseCapture();
+    dragidx = -1;
+    SetFocus(wnd);
   }
   else if (msg == WM_NOTIFY) {
     //printf("tabbar con_proc WM_NOTIFY\n");
@@ -305,12 +289,6 @@ static int targpro = -1;
         if (tabinfo[i].wnd == wnd)
           SendMessage(tab_wnd, TCM_SETCURSEL, i, 0);
       }
-    }
-    else if (lpnmhdr->code == (uint)NM_CLICK) {
-      // clear tab dragging; the tab drop could be hooked here 
-      // but would not work if dropped in free space right of tabset
-      dragmsg = 0;
-      dragidx = -1;
     }
   }
   else if (msg == WM_CREATE) {
@@ -368,12 +346,39 @@ static int targpro = -1;
       SetTextColor(hdc, GetSysColor(COLOR_HIGHLIGHTTEXT));
       //printf("TAB fg %06X\n", GetSysColor(COLOR_HIGHLIGHTTEXT));
 
-      // override active tab colours if configured
-      tabbg = cfg.tab_bg_colour;
-      if (tabbg != (colour)-1) {
-        //printf("TAB bg %06X\n", tabbg);
+      // override active tab colours
+      if (dragidx >= 0 && dragidx != dragini) {
+        // dynamic highlighting of tab being dragged (not initially)
+        colour bg1 = cfg.tab_bg_colour;
+        if (bg1 == (colour)-1)
+          bg1 = GetSysColor(COLOR_HIGHLIGHT);
+        colour bg0 = GetSysColor(COLOR_3DFACE);
+
+        //int p = targpro * 80 / 100 + 10;
+        int p = (100 - sqr(100 - targpro) / 100) * 80 / 100 + 10;
+        int r = red(bg0) + p * (red(bg1) - red(bg0)) / 100;
+        int g = green(bg0) + p * (green(bg1) - green(bg0)) / 100;
+        int b = blue(bg0) + p * (blue(bg1) - blue(bg0)) / 100;
+
+        // set dynamic highlighting colour, depending on tab middle focus
+        tabbg = RGB(r, g, b);
+        // could use a different colour, or even configurable:
+        //tabbg = RGB(b, g, r);
+        //tabbg = RGB(g, b, r);
+        //tabbg = RGB(g, r, b);
+        //tabbg = RGB(b, r, g);
+
         tabbr = CreateSolidBrush(tabbg);
       }
+      else {
+        // override active tab colours if configured
+        tabbg = cfg.tab_bg_colour;
+        if (tabbg != (colour)-1) {
+          //printf("TAB bg %06X\n", tabbg);
+          tabbr = CreateSolidBrush(tabbg);
+        }
+      }
+      // foreground colour
       colour tabfg = cfg.tab_fg_colour;
       if (tabfg != (colour)-1) {
         //printf("TAB fg %06X\n", tabfg);
@@ -386,23 +391,6 @@ static int targpro = -1;
       //tabbr = GetSysColorBrush(COLOR_INACTIVECAPTION);
       SetTextColor(hdc, GetSysColor(COLOR_CAPTIONTEXT));
       //printf("tab fg %06X\n", GetSysColor(COLOR_CAPTIONTEXT));
-
-      // drag-and-drop hover highlighting
-      if (itemID == targidx) {
-        colour bg1 = cfg.tab_bg_colour;
-        if (bg1 == (colour)-1)
-          bg1 = GetSysColor(COLOR_HIGHLIGHT);
-        colour bg0 = GetSysColor(COLOR_3DFACE);
-
-        //int p = targpro * 80 / 100 + 10;
-        int p = (100 - sqr(100 - targpro) / 100) * 80 / 100 + 10;
-        int r = red(bg0) + p * (red(bg1) - red(bg0)) / 100;
-        int g = green(bg0) + p * (green(bg1) - green(bg0)) / 100;
-        int b = blue(bg0) + p * (blue(bg1) - blue(bg0)) / 100;
-
-        tabbg = RGB(r, g, b);
-        tabbr = CreateSolidBrush(tabbg);
-      }
     }
     FillRect(hdc, &dis->rcItem, tabbr);
     if (tabbg != (colour)-1)