+++ wiki/Changelog.md
@@ -1,9 +1,13 @@
 Character rendering
   * Reenabled distinct blinking for combined characters (broken since 3.5.3).
+  * Fix skipping of accent to next position on background colour.
   * Fix drawing of Powerline round symbols (mintty/wsltty#358).
 
 Window handling
   * Show new tab at its target position right away (#1292).
+  * Live changing of reordered tabs while being dragged (#1293).
+  * Change mouse pointer while tab dragging (#1293).
+  * Adapt dynamic tab drag highlighting (~#1293).
 
 ### 3.7.6 (24 September 2024) ###